requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
APScheduler>=3.10.0,<4.0.0
psycopg2-binary>=2.9.0
